+++ lib/git/src/checkout.dart
