+++ lib/git/src/exceptions/no_commits_exception.dart
@@ -1,0 +1,1 @@
+class NoCommitsException implements Exception {}
