+++ lib/git/src/models/commit.dart
@@ -1,16 +1,16 @@
 import 'package:equatable/equatable.dart';
 
-class GitCommit extends Equatable {
+class Commit extends Equatable {
   final String hash;
   final DateTime date;
 
-  GitCommit(this.hash, this.date);
+  Commit(this.hash, this.date);
 
   @override
   List<Object> get props => [hash, date];
 }
 
-GitCommit commitFromStdOut(String stdOut) {
+Commit commitFromStdOut(String stdOut) {
   var hashAndDate = stdOut.split('"')[1].split(', ');
-  return GitCommit(hashAndDate[0], DateTime.parse(hashAndDate[1]));
+  return Commit(hashAndDate[0], DateTime.parse(hashAndDate[1]));
 }
