+++ lib/git/src/version.dart
