+++ lib/requirements_checker.dart
@@ -1,13 +1,13 @@
 import 'dart:io';
 
-import 'package:peot/git/git_version.dart' as git_version;
+import 'git/git.dart' as git;
 
 import 'composer.dart' as composer;
 
 Future<void> checkRequirements() async {
   print('checking requirements...');
   try {
-    await git_version.version();
+    await git.version();
     await composer.version();
   } on ProcessException catch (e) {
     throw MissingRequirementException(e);
